[tech] extract 'enhance_pickup_dropoff' as an enhancer sub-module
