+++ src/enhancers/mod.rs
@@ -1,7 +1,9 @@
 //! This module contains various functions that enhance / cleanup `Collections`
 
 mod adjust_lines_names;
+mod enhance_pickup_dropoff;
 mod fill_co2;
 
 pub(crate) use adjust_lines_names::adjust_lines_names;
+pub(crate) use enhance_pickup_dropoff::enhance_pickup_dropoff;
 pub(crate) use fill_co2::fill_co2;
